refactor: 루트 레이아웃에서 Footer 제거 및 Footer를 사용하는 폴더 구조 따로 구성 (#96)
