+++ src/app/(with-footer)/faq/_components/FaqAccordion.tsx
