+++ src/app/(with-footer)/recruit/_components/ActivityCard.tsx
