+++ src/app/(with-footer)/recruit/_components/PositionCard.tsx
