+++ src/app/(with-footer)/recruit/_components/QualificationsCard.tsx
